Remove old xldb analyzers and create separate xldbanalyzer with new dll
